+++ RandomSkunk.Results/Result{T}.cs
@@ -18,13 +18,13 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
         _error = null;
     }
 
-    private Result(Error? error, bool? setStackTrace)
+    private Result(Error? error, bool? omitStackTrace)
     {
         _outcome = Outcome.Fail;
         _value = default;
         _error = error ?? new Error();
 
-        if (_error.StackTrace is null && (setStackTrace ?? FailResult.SetStackTrace))
+        if (_error.StackTrace is null && !(omitStackTrace ?? FailResult.OmitStackTrace))
             _error = _error with { StackTrace = FilteredStackTrace.Create() };
 
         _error = FailResult.InvokeReplaceErrorIfSet(_error);
@@ -109,12 +109,12 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
     /// Creates a <c>Fail</c> result with the specified error.
     /// </summary>
     /// <param name="error">An error that describes the failure. If <see langword="null"/>, a default error is used.</param>
-    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
-    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
-    ///     instead to determine whether to set the stack trace.</param>
+    /// <param name="omitStackTrace">Whether to omit the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.OmitStackTrace"/> property is used
+    ///     instead to determine whether to omit the stack trace.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
-    public static Result<T> Fail(Error? error = null, bool? setStackTrace = null) => new(error, setStackTrace);
+    public static Result<T> Fail(Error? error = null, bool? omitStackTrace = null) => new(error, omitStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -125,9 +125,9 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
     /// <param name="errorIdentifier">The optional identifier of the error.</param>
     /// <param name="errorTitle">The optional title for the error. If <see langword="null"/>, then "Error" is used instead.
     ///     </param>
-    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
-    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
-    ///     instead to determine whether to set the stack trace.</param>
+    /// <param name="omitStackTrace">Whether to omit the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.OmitStackTrace"/> property is used
+    ///     instead to determine whether to omit the stack trace.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Result<T> Fail(
@@ -136,8 +136,8 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
         int? errorCode = ErrorCodes.CaughtException,
         string? errorIdentifier = null,
         string? errorTitle = null,
-        bool? setStackTrace = null) =>
-        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle), setStackTrace);
+        bool? omitStackTrace = null) =>
+        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle), omitStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -150,9 +150,9 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
     /// <param name="isSensitive">Whether the error contains sensitive information.</param>
     /// <param name="extensions">Additional properties for the error.</param>
     /// <param name="innerError">The optional error that is the cause of the current error.</param>
-    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
-    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
-    ///     instead to determine whether to set the stack trace.</param>
+    /// <param name="omitStackTrace">Whether to omit the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.OmitStackTrace"/> property is used
+    ///     instead to determine whether to omit the stack trace.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Result<T> Fail(
@@ -163,7 +163,7 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
         bool isSensitive = false,
         IReadOnlyDictionary<string, object>? extensions = null,
         Error? innerError = null,
-        bool? setStackTrace = null) =>
+        bool? omitStackTrace = null) =>
         Fail(
             new Error
             {
@@ -175,7 +175,7 @@ public readonly partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
                 Extensions = extensions!,
                 InnerError = innerError,
             },
-            setStackTrace);
+            omitStackTrace);
 
     /// <summary>
     /// Creates a <c>Success</c> result with the specified value. If the value is <see langword="null"/>, then a <c>Fail</c>