+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning].
     - Change `Maybe.None()` method to property.
     - Rename `ErrorCodes.NoneResult` to `ErrorCodes.NoValue`.
     - Replace `DBNull` with new `Unit` type.
+    - Rename `SetStackTrace` to `OmitStackTrace` and flip logic.
 
 ### Added